feat: scan result page to show result counts. - detailed result ui is wip
