+++ mobileapp/src/pages/Accounts/ScanResult/index.tsx
@@ -15,46 +15,23 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import { Button, Divider, List, ListItem, Text } from '@ui-kitten/components';
-import React, { useEffect } from 'react';
-import { Pressable, View, ScrollView, Image } from 'react-native';
-import Layout from '../../components/Layout';
-
-import { StyleService, useStyleSheet } from '@ui-kitten/components';
-import ConnectedApps from '../../selectors/google/ConnectedApps';
+import { Divider, StyleService, Text, useStyleSheet } from '@ui-kitten/components';
+import React from 'react';
+import { Pressable, View } from 'react-native';
 
 const themedStyles = StyleService.create({
-  card: {
-    flex: 1,
-    minHeight: 130,
-    marginVertical: 10,
-    marginHorizontal: 20,
-    flexDirection: 'row',
-    backgroundColor: 'white',
-    borderRadius: 8,
-    shadowColor: '#000',
-    shadowOffset: {
-      width: 0,
-      height: 1,
-    },
-    shadowOpacity: 0.22,
-    shadowRadius: 2.22,
-  },
   root: {
     flex: 1,
-    // marginVertical: 50,
-    marginHorizontal: 14,
+    marginHorizontal: 5,
     flexDirection: 'row',
     justifyContent: 'space-between',
     alignItems: 'center',
-    flexWrap: 'wrap',
   },
-  iconButton: {
+  smallCard: {
     flex: 1,
     minHeight: 120,
     minWidth: 100,
     maxHeight: 120,
-    // maxWidth: 100,
     marginHorizontal: 7,
     marginVertical: 10,
     backgroundColor: 'white',
@@ -68,73 +45,31 @@ const themedStyles = StyleService.create({
     shadowRadius: 2.22,
     elevation: 3,
   },
-  icon: {
-    // marginHorizontal: 4,
-    // marginVertical: 4,
-    alignSelf: 'center',
-    height: 55,
-    width: 55,
+  issueBold: {
+    color: 'red',
+    flex: 4,
+    fontSize: 60,
+    alignSelf: 'flex-start',
+  },
+  text: {
+    flex: 4,
+    fontSize: 20,
+    marginVertical: 10,
+    paddingLeft: 3,
   },
   touch: {
     flex: 1,
     flexDirection: 'row',
-    padding: 10,
-    flexWrap: 'wrap',
-  },
-  issueBold: {
-    color: 'red',
-    flex: 2,
-    fontSize: 80,
-    alignSelf: 'flex-start',
-  },
-  text: {
-    flex: 2,
-    fontSize: 19,
-    flexWrap: 'wrap',
-    marginVertical: 15,
-  },
-  headerText: {
-    flex: 2,
-    marginHorizontal: 15,
-    fontSize: 19,
-    flexWrap: 'wrap',
-    marginVertical: 15,
-  },
-  button: {
-    flex: 1,
-    justifyContent: 'flex-start',
-    padding: 0,
-    alignSelf: 'flex-start',
-  },
-
-  desc: {
-    flex: 4,
-    padding: 24,
-    alignItems: 'flex-start',
-    justifyContent: 'flex-start',
-  },
-
-  imageContainer: {
-    flex: 2,
-  },
-  image: {
-    flex: 1,
-    height: '100%',
-    width: '100%',
-    // resizeMode: 'contain',
+    paddingLeft: 5,
+    paddingTop: 15,
   },
 
   connectedAccCard: {
     flex: 1,
-    padding: 10,
-  },
-  tinyBox: {
-    flex: 1,
-    flexDirection: 'column',
-    minHeight: 120,
-    minWidth: 100,
-    marginHorizontal: 5,
-    marginVertical: 20,
+    flexDirection: 'row',
+    padding: 5,
+    marginHorizontal: 20,
+    marginVertical: 10,
     backgroundColor: 'white',
     borderRadius: 8,
     shadowColor: '#000',
@@ -144,19 +79,13 @@ const themedStyles = StyleService.create({
     },
     shadowOpacity: 0.22,
     shadowRadius: 2.22,
+    elevation: 3,
   },
-  deviceLogo: {
+  connectedAccBoldText: {
+    color: 'gold',
     flex: 2,
-    width: 80,
-    height: 50,
-  },
-  tinyText: {
-    flex: 1,
-  },
-  appLogo: {
-    flex: 2,
-    width: 50,
-    height: 50,
+    fontSize: 70,
+    alignSelf: 'flex-start',
   },
 });
 
@@ -164,38 +93,34 @@ type accountProps = {
   navigation: any;
 };
 
-export default function ScanResults(props: any) {
+export default function ScanResultSummary(props: any) {
   const styles = useStyleSheet(themedStyles);
 
-  const { connectedDevices, thirdPartyApps } = props.result;
+  const {
+    connectedDevices,
+    thirdPartyApps,
+    privacyIssuesCount,
+    securityIssuesCount,
+  } = props.result;
 
-  function getDevices(dType: string, imgURL: string) {
-    switch (dType) {
-      case 'android':
-        return imgURL;
-      case 'Linux':
-        return 'https://storage.googleapis.com/fireser-app-public-assets/icons/devices/linuxlaptop.png';
-      case 'iphone':
-        return 'https://storage.googleapis.com/fireser-app-public-assets/icons/devices/iphone-x-icon.png';
-      case 'Mac':
-        return 'https://storage.googleapis.com/fireser-app-public-assets/icons/devices/mac.png';
-      case 'Windows':
-        return 'https://storage.googleapis.com/fireser-app-public-assets/icons/devices/winlaptop.png';
-      default:
-        return 'https://www.gstatic.com/identity/boq/accountsettingssecuritycommon/images/sprites/devices_realistic_72-ef7d73f742f5343ba1bd3c1e6b13ffba.png';
+  // formatNumber adds 0 to single digit number
+  function formatNumber(num: number) {
+    if (num.toString().length === 1) {
+      return '0' + num.toString();
+    } else {
+      return num;
     }
   }
-
   return (
     <View>
       <View style={styles.root}>
-        <View style={styles.iconButton}>
+        <View style={styles.smallCard}>
           <Pressable
             // onPress={() => navigation.navigate('ScanAndProtect', { name: 'Google' })}
             style={styles.touch}
           >
             <Text style={styles.issueBold} category="h1">
-              4
+              {formatNumber(securityIssuesCount)}
             </Text>
 
             <Text category="s1" style={styles.text}>
@@ -204,13 +129,13 @@ export default function ScanResults(props: any) {
           </Pressable>
         </View>
         <Divider />
-        <View style={styles.iconButton}>
+        <View style={styles.smallCard}>
           <Pressable
             // onPress={() => navigation.navigate('ScanAndProtect', { name: 'Google' })}
             style={styles.touch}
           >
             <Text style={styles.issueBold} category="h1">
-              3
+              {formatNumber(privacyIssuesCount)}
             </Text>
 
             <Text category="s1" style={styles.text}>
@@ -221,75 +146,23 @@ export default function ScanResults(props: any) {
       </View>
 
       <View style={styles.connectedAccCard}>
-        <Text category="s1" style={styles.headerText}>
-          {connectedDevices?.length} devices connected to this account{' '}
+        <Text style={styles.connectedAccBoldText} category="h1">
+          {formatNumber(connectedDevices?.length)}
         </Text>
 
-        <ScrollView horizontal={true}>
-          {connectedDevices.map((d: any, i: number) => (
-            <View key={i} style={styles.tinyBox}>
-              <Pressable
-              // onPress={() => navigation.navigate('ScanAndProtect', { name: 'Google' })}
-              // style={styles.touch}
-              >
-                <Image
-                  style={styles.deviceLogo}
-                  source={{
-                    uri: d.imgURL ? d.imgURL : getDevices(d.title, d.imgURL),
-                  }}
-                />
-
-                <Text category="s1" style={styles.tinyText}>
-                  {d.title}
-                </Text>
-              </Pressable>
-            </View>
-          ))}
-        </ScrollView>
+        <Text category="s1" style={styles.text}>
+          Devices are connected to your account{' '}
+        </Text>
       </View>
 
       <View style={styles.connectedAccCard}>
-        <Text category="s1" style={styles.headerText}>
-          {thirdPartyApps?.length} 3rd party apps have access to your data{' '}
+        <Text style={styles.connectedAccBoldText} category="h1">
+          {formatNumber(thirdPartyApps?.length)}
         </Text>
-        <ScrollView horizontal={true}>
-          {thirdPartyApps?.map((d: any, i: number) => (
-            <View style={styles.tinyBox} key={i}>
-              <Pressable>
-                <Image
-                  style={styles.appLogo}
-                  source={{
-                    uri: d.imgURL,
-                  }}
-                />
-
-                <Text category="s1" style={styles.tinyText}>
-                  {d.name}
-                </Text>
-              </Pressable>
-            </View>
-          ))}
-        </ScrollView>
+        <Text category="s1" style={styles.text}>
+          Third party apps have access to your data{' '}
+        </Text>
       </View>
-
-      {/* <View style={styles.card}>
-         <Pressable onPress={props.navigateToAccounts} style={styles.touch}>
-           <View style={styles.desc}>
-             <Text category="h5">Your Public Profile</Text>
-             <Text category="s1">See how your private data represents you</Text>
-             <Button
-               style={styles.button}
-               appearance="ghost"
-               status="primary"
-               onPress={props.navigateToAccounts}
-             >
-               Check it now
-             </Button>
-           </View>
- 
-           <View style={styles.imageContainer}><AccountSec style={styles.image} /></View>
-         </Pressable>
-       </View> */}
     </View>
   );
 }
